pyproject.toml

[project]
name = "agent-outreach-platform"
version = "1.0.0"
description = "Agent-driven outreach and qualification platform"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "gradio>=5.22.0",
    "openai-agents>=0.13.2",
    "agentmail",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic-settings>=2.0.0",
    "composio-openai-agents>=0.5.0",
    "composio>=0.5.0",
    "composio-openai>=0.11.4",
    "tenacity>=9.1.4",
    "python-jose[cryptography]>=3.5.0",
    "httpx>=0.28.1",
]

[project.scripts]
run-email-monitor = "run_email_monitor:main"
setup-webhook = "setup_webhook:main"
test-email-monitor = "test_email_monitor:main"

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
